Add warning api
The idea is that a Warning is essentially a "builder" for a warning - we
apply changes to it (e.g. adding strings) and eventually output it with
warn_finish.

Doing it this way allows us to add more information to warnings later,
such as the filename & line number. Having individual tt_warn_t objects
instead of a global warning allows multiple warnings to be constructed
simultaneously (e.g. across multiple functions).

For now, I'm putting the warning functions in the api bridge. This only
necessary for warn_finish, which requires the context to print the
message. I'm not sure where the others belong - there should be a way to
import Rust functions, but I'm not sure how to do this.